scripts: Use destdir_join fix for all DESTDIR prefixing
Also fixes a bug in the gettext module where an exception would be
thrown if DESTDIR wasn't exported in the environment
